Move filter_jobs_on_mode to utils
